Fix error

+++ lan_sync_controller/base.py
@@ -91,7 +91,8 @@ class BaseDaemon(object):
         open(self.pidfile, 'w+').write("%s\n" % pid)
 
     def delpid(self):
-        os.remove(self.pidfile)
+        if os.path.exists(self.pidfile):
+            os.remove(self.pidfile)
 
     def status(self):
         try:
